+++ src/main/kotlin/model/judge/ThreeThreeCheck.kt
@@ -1,4 +1,9 @@
-package model
+package model.judge
+
+import model.Direction
+import model.Position
+import model.Stone
+import model.StoneColor
 
 object ThreeThreeCheck : Rule {
     override fun checkFoulByAllDirections(
